✨ feat: refactor vue template
