+++ internal/plugins/frameworks/vue.go
@@ -2,10 +2,15 @@ package frameworks
 
 import (
 	"path/filepath"
+	"strings"
+
+	tilocontext "tilokit/internal/core/context"
+	"tilokit/internal/templates/common"
+	"tilokit/internal/templates/vue"
+	"tilokit/internal/utils"
 
 	"github.com/pkg/errors"
-	tilocontext "tilokit/internal/core/context"
-	"tilokit/internal/utils"
+
 	"tilokit/pkg/constants"
 )
 
@@ -93,614 +98,110 @@ func (p *VuePlugin) createDirectoryStructure(ctx *tilocontext.ExecutionContext) 
 }
 
 func (p *VuePlugin) generatePackageJson(ctx *tilocontext.ExecutionContext) error {
-	packageJson := `{
-  "name": "` + ctx.Config.ProjectName + `",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
-    "preview": "vite preview",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
-    "type-check": "vue-tsc --noEmit"
-  },
-  "dependencies": {
-    "vue": "^3.4.0",
-    "vue-router": "^4.2.5",
-    "pinia": "^2.1.7"
-  },
-  "devDependencies": {
-    "@rushstack/eslint-patch": "^1.3.3",
-    "@tsconfig/node18": "^18.2.2",
-    "@types/node": "^18.18.5",
-    "@vitejs/plugin-vue": "^4.4.0",
-    "@vue/eslint-config-prettier": "^8.0.0",
-    "@vue/eslint-config-typescript": "^12.0.0",
-    "@vue/tsconfig": "^0.4.0",
-    "eslint": "^8.49.0",
-    "eslint-plugin-vue": "^9.17.0",
-    "npm-run-all2": "^6.1.1",
-    "prettier": "^3.0.3",
-    "typescript": "~5.2.0",
-    "vite": "^5.0.0",
-    "vue-tsc": "^1.8.19"
-  }
-}`
+	packageJson := vue.ViteTsPackageJson
+	packageJson = strings.ReplaceAll(packageJson, "{{.ProjectName}}", ctx.Config.ProjectName)
 
 	packageJsonPath := filepath.Join(ctx.ProjectPath, "package.json")
 	return utils.WriteFile(packageJsonPath, packageJson)
 }
 
 func (p *VuePlugin) generateSourceFiles(ctx *tilocontext.ExecutionContext) error {
-	// Generate main.ts
-	mainTs := `import { createApp } from 'vue'
-import { createPinia } from 'pinia'
-import App from './App.vue'
-import router from './router'
-
-import './assets/main.css'
-
-const app = createApp(App)
-
-app.use(createPinia())
-app.use(router)
-
-app.mount('#app')`
-
-	// Generate App.vue
-	appVue := `<script setup lang="ts">
-import { RouterLink, RouterView } from 'vue-router'
-import HelloWorld from './components/HelloWorld.vue'
-</script>
-
-<template>
-  <header>
-    <div class="wrapper">
-      <HelloWorld msg="Welcome to ` + ctx.Config.ProjectName + `" />
-
-      <nav>
-        <RouterLink to="/">Home</RouterLink>
-        <RouterLink to="/about">About</RouterLink>
-      </nav>
-    </div>
-  </header>
-
-  <RouterView />
-</template>
-
-<style scoped>
-header {
-  line-height: 1.5;
-  max-height: 100vh;
-}
-
-nav {
-  width: 100%;
-  font-size: 12px;
-  text-align: center;
-  margin-top: 2rem;
-}
-
-nav a.router-link-exact-active {
-  color: var(--color-text);
-}
-
-nav a.router-link-exact-active:hover {
-  background-color: transparent;
-}
-
-nav a {
-  display: inline-block;
-  padding: 0 1rem;
-  border-left: 1px solid var(--color-border);
-}
-
-nav a:first-of-type {
-  border: 0;
-}
-
-@media (min-width: 1024px) {
-  header {
-    display: flex;
-    place-items: center;
-    padding-right: calc(var(--section-gap) / 2);
-  }
-
-  .logo {
-    margin: 0 2rem 0 0;
-  }
-
-  header .wrapper {
-    display: flex;
-    place-items: flex-start;
-    flex-wrap: wrap;
-  }
-
-  nav {
-    text-align: left;
-    margin-left: -1rem;
-    font-size: 1rem;
-
-    padding: 1rem 0;
-    margin-top: 1rem;
-  }
-}
-</style>`
-
-	// Generate HelloWorld.vue
-	helloWorldVue := `<script setup lang="ts">
-defineProps<{
-  msg: string
-}>()
-</script>
-
-<template>
-  <div class="greetings">
-    <h1 class="green">{{ msg }}</h1>
-    <h3>
-      You've successfully created a project with
-      <a href="https://vitejs.dev/" target="_blank" rel="noopener">Vite</a> +
-      <a href="https://vuejs.org/" target="_blank" rel="noopener">Vue 3</a> +
-      <a href="https://www.typescriptlang.org/" target="_blank" rel="noopener">TypeScript</a>.
-    </h3>
-  </div>
-</template>
-
-<style scoped>
-h1 {
-  font-weight: 500;
-  font-size: 2.6rem;
-  position: relative;
-  top: -10px;
-}
-
-h3 {
-  font-size: 1.2rem;
-}
-
-.greetings h1,
-.greetings h3 {
-  text-align: center;
-}
-
-@media (min-width: 1024px) {
-  .greetings h1,
-  .greetings h3 {
-    text-align: left;
-  }
-}
-</style>`
-
-	// Generate router
-	routerTs := `import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
-
-const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: '/',
-      name: 'home',
-      component: HomeView
-    },
-    {
-      path: '/about',
-      name: 'about',
-      component: () => import('../views/AboutView.vue')
-    }
-  ]
-})
-
-export default router`
-
-	// Generate views
-	homeViewVue := `<script setup lang="ts">
-import TheWelcome from '../components/TheWelcome.vue'
-</script>
-
-<template>
-  <main>
-    <TheWelcome />
-  </main>
-</template>`
-
-	aboutViewVue := `<template>
-  <div class="about">
-    <h1>This is an about page</h1>
-  </div>
-</template>
-
-<style>
-@media (min-width: 1024px) {
-  .about {
-    min-height: 100vh;
-    display: flex;
-    align-items: center;
-  }
-}
-</style>`
-
-	theWelcomeVue := `<script setup lang="ts">
-import WelcomeItem from './WelcomeItem.vue'
-import DocumentationIcon from './icons/IconDocumentation.vue'
-import ToolingIcon from './icons/IconTooling.vue'
-import EcosystemIcon from './icons/IconEcosystem.vue'
-import CommunityIcon from './icons/IconCommunity.vue'
-import SupportIcon from './icons/IconSupport.vue'
-</script>
-
-<template>
-  <WelcomeItem>
-    <template #icon>
-      <DocumentationIcon />
-    </template>
-    <template #heading>Documentation</template>
-
-    Vue's
-    <a href="https://vuejs.org/" target="_blank" rel="noopener">official documentation</a>
-    provides you with all information you need to get started.
-  </WelcomeItem>
-
-  <WelcomeItem>
-    <template #icon>
-      <ToolingIcon />
-    </template>
-    <template #heading>Tooling</template>
-
-    This project is served and bundled with
-    <a href="https://vitejs.dev/guide/features.html" target="_blank" rel="noopener">Vite</a>. The
-    recommended IDE setup is
-    <a href="https://code.visualstudio.com/" target="_blank" rel="noopener">VSCode</a> +
-    <a href="https://github.com/johnsoncodehk/volar" target="_blank" rel="noopener">Volar</a>. If
-    you need to test your components and web pages, check out
-    <a href="https://www.cypress.io/" target="_blank" rel="noopener">Cypress</a> and
-    <a href="https://on.cypress.io/component" target="_blank" rel="noopener"
-      >Cypress Component Testing</a
-    >.
-  </WelcomeItem>
-
-  <WelcomeItem>
-    <template #icon>
-      <EcosystemIcon />
-    </template>
-    <template #heading>Ecosystem</template>
-
-    Get official tools and libraries for your project:
-    <a href="https://pinia.vuejs.org/" target="_blank" rel="noopener">Pinia</a>,
-    <a href="https://router.vuejs.org/" target="_blank" rel="noopener">Vue Router</a>,
-    <a href="https://test-utils.vuejs.org/" target="_blank" rel="noopener">Vue Test Utils</a>, and
-    <a href="https://github.com/vuejs/devtools" target="_blank" rel="noopener">Vue Dev Tools</a>. If
-    you need more resources, we suggest paying
-    <a href="https://github.com/vuejs/awesome-vue" target="_blank" rel="noopener">Awesome Vue</a>
-    a visit.
-  </WelcomeItem>
-
-  <WelcomeItem>
-    <template #icon>
-      <CommunityIcon />
-    </template>
-    <template #heading>Community</template>
-
-    Got stuck? Ask your question on
-    <a href="https://chat.vuejs.org" target="_blank" rel="noopener">Vue Land</a>, our official
-    Discord server, or
-    <a href="https://stackoverflow.com/questions/tagged/vue.js" target="_blank" rel="noopener"
-      >StackOverflow</a
-    >. You should also subscribe to
-    <a href="https://news.vuejs.org" target="_blank" rel="noopener">our mailing list</a> and follow
-    <a href="https://twitter.com/vuejs" target="_blank" rel="noopener">the official Vue Twitter account</a>
-    for latest news in the Vue world.
-  </WelcomeItem>
-
-  <WelcomeItem>
-    <template #icon>
-      <SupportIcon />
-    </template>
-    <template #heading>Support Vue</template>
-
-    As an independent project, Vue relies on community backing for its sustainability. You can help
-    us by
-    <a href="https://vuejs.org/sponsor/" target="_blank" rel="noopener">becoming a sponsor</a>.
-  </WelcomeItem>
-</template>`
-
-	// Generate CSS
-	mainCss := `@import './base.css';
-
-#app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-
-  font-weight: normal;
-}
-
-a,
-.green {
-  text-decoration: none;
-  color: hsla(160, 100%, 37%, 1);
-  transition: 0.4s;
-}
-
-@media (hover: hover) {
-  a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
-  }
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    padding: 0 2rem;
-  }
-}`
-
-	// Generate base.css
-	baseCss := `/* CSS Reset and Base Styles */
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-  margin: 0;
-  font-weight: normal;
-}
-
-body {
-  min-height: 100vh;
-  color: var(--color-text);
-  background: var(--color-background);
-  transition:
-    color 0.5s,
-    background-color 0.5s;
-  line-height: 1.6;
-  font-family:
-    Inter,
-    -apple-system,
-    BlinkMacSystemFont,
-    'Segoe UI',
-    Roboto,
-    Oxygen,
-    Ubuntu,
-    Cantarell,
-    'Fira Sans',
-    'Droid Sans',
-    'Helvetica Neue',
-    sans-serif;
-  font-size: 15px;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-/* CSS Variables */
-:root {
-  --color-background: #ffffff;
-  --color-background-soft: #f8f8f8;
-  --color-background-mute: #f2f2f2;
-  --color-border: #d1d5db;
-  --color-border-hover: #b1b5bb;
-  --color-heading: #2c3e50;
-  --color-text: #213547;
-  --section-gap: 160px;
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --color-background: #1a1a1a;
-    --color-background-soft: #242424;
-    --color-background-mute: #2c2c2c;
-    --color-border: #3c3c3c;
-    --color-border-hover: #5c5c5c;
-    --color-heading: #ffffff;
-    --color-text: #ebebeb;
-  }
-}`
-
-	// Generate WelcomeItem.vue
-	welcomeItemVue := `<template>
-  <div class="item">
-    <i>
-      <slot name="icon"></slot>
-    </i>
-    <div class="details">
-      <h3>
-        <slot name="heading"></slot>
-      </h3>
-      <slot></slot>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.item {
-  margin-top: 2rem;
-  display: flex;
-}
-
-.details {
-  flex: 1;
-  margin-left: 1rem;
-}
-
-i {
-  display: flex;
-  place-items: center;
-  place-content: center;
-  width: 32px;
-  height: 32px;
-  color: var(--color-text);
-}
-
-h3 {
-  font-size: 1.2rem;
-  font-weight: 500;
-  margin-bottom: 0.4rem;
-  color: var(--color-heading);
-}
-
-@media (min-width: 1024px) {
-  .item {
-    margin-top: 0;
-    padding: 0.4rem 0 1rem calc(var(--section-gap) / 2);
-  }
-
-  i {
-    top: calc(50% - 25px);
-    left: -26px;
-    position: absolute;
-    border: 1px solid var(--color-border);
-    background: var(--color-background);
-    border-radius: 8px;
-    width: 50px;
-    height: 50px;
-  }
-
-  .item:before {
-    content: ' ';
-    border-left: 1px solid var(--color-border);
-    position: absolute;
-    left: 0;
-    bottom: calc(50% + 25px);
-    height: calc(50% - 25px);
-  }
-
-  .item:after {
-    content: ' ';
-    border-left: 1px solid var(--color-border);
-    position: absolute;
-    left: 0;
-    top: calc(50% + 25px);
-    height: calc(50% - 25px);
-  }
-
-  .item:first-of-type:before {
-    display: none;
-  }
-
-  .item:last-of-type:after {
-    display: none;
-  }
-}
-</style>`
-
-	// Generate simple icons (SVG-based)
-	iconDocumentation := `<template>
-  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="17" fill="currentColor">
-    <path d="M11 2.253a1 1 0 1 0-2 0h2zm-2 13a1 1 0 1 0 2 0H9zM9 2v13h2V2H9z"/>
-    <path d="M5 3a2 2 0 0 0-2 2v6a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V5a2 2 0 0 0-2-2H5zM3 5a2 2 0 0 1 2-2h6a2 2 0 0 1 2 2v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V5z"/>
-  </svg>
-</template>`
-
-	iconTooling := `<template>
-  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor">
-    <path d="M8 4a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v4H8V4zM6 8h8v8a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2V8z"/>
-  </svg>
-</template>`
-
-	iconEcosystem := `<template>
-  <svg xmlns="http://www.w3.org/2000/svg" width="18" height="20" fill="currentColor">
-    <path d="M9 0a9 9 0 1 0 9 9A9 9 0 0 0 9 0zm4.5 7.5h-9v-3h9v3zm0 1.5v3h-9v-3h9z"/>
-  </svg>
-</template>`
-
-	iconCommunity := `<template>
-  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor">
-    <path d="M15 4a1 1 0 1 0 0 2V4zM2 9a1 1 0 0 0 0 2V9zm13-5a1 1 0 1 0 0 2V4zM5 9a1 1 0 0 0 0 2V9z"/>
-  </svg>
-</template>`
-
-	iconSupport := `<template>
-  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor">
-    <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5zM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3zM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2z"/>
-  </svg>
-</template>`
-
-	// Write files
-	files := map[string]string{
-		"src/main.ts":                                mainTs,
-		"src/App.vue":                                appVue,
-		"src/components/HelloWorld.vue":              helloWorldVue,
-		"src/components/WelcomeItem.vue":             welcomeItemVue,
-		"src/components/icons/IconDocumentation.vue": iconDocumentation,
-		"src/components/icons/IconTooling.vue":       iconTooling,
-		"src/components/icons/IconEcosystem.vue":     iconEcosystem,
-		"src/components/icons/IconCommunity.vue":     iconCommunity,
-		"src/components/icons/IconSupport.vue":       iconSupport,
-		"src/router/index.ts":                        routerTs,
-		"src/views/HomeView.vue":                     homeViewVue,
-		"src/views/AboutView.vue":                    aboutViewVue,
-		"src/components/TheWelcome.vue":              theWelcomeVue,
-		"src/assets/main.css":                        mainCss,
-		"src/assets/base.css":                        baseCss,
+	language := ctx.Variables["language"].(string)
+	isJS := language == "js"
+
+	// Define file templates based on language
+	var templates map[string]string
+	if isJS {
+		templates = p.getJavaScriptTemplates()
+	} else {
+		templates = p.getTypeScriptTemplates()
 	}
 
-	for path, content := range files {
-		fullPath := filepath.Join(ctx.ProjectPath, path)
+	// Write all files
+	for relativePath, content := range templates {
+		fullPath := filepath.Join(ctx.ProjectPath, relativePath)
 		if err := utils.WriteFile(fullPath, content); err != nil {
 			return err
 		}
 	}
 
 	return nil
+}
+
+func (p *VuePlugin) getJavaScriptTemplates() map[string]string {
+	return map[string]string{
+		"src/main.js":                                vue.ViteJsMainFile,
+		"src/App.vue":                                vue.ViteJsAppFile,
+		"src/components/HelloWorld.vue":              vue.ViteHelloWordVue,
+		"src/components/WelcomeItem.vue":             vue.ViteJsWelcomeItem,
+		"src/components/TheWelcome.vue":              vue.ViteJsTheWelcome,
+		"src/components/icons/IconDocumentation.vue": vue.ViteIconDocumentation,
+		"src/components/icons/IconTooling.vue":       vue.ViteIconTooling,
+		"src/components/icons/IconEcosystem.vue":     vue.ViteIconEcoSystem,
+		"src/components/icons/IconCommunity.vue":     vue.ViteIconCommunity,
+		"src/components/icons/IconSupport.vue":       vue.ViteIconSupport,
+		"src/router/index.js":                        vue.ViteJsRouter,
+		"src/views/HomeView.vue":                     vue.ViteJsHomeView,
+		"src/views/AboutView.vue":                    vue.ViteJsAboutView,
+		"src/assets/main.css":                        vue.ViteJsMainCss,
+		"src/assets/base.css":                        vue.ViteJsBaseCss,
+	}
+}
+
+func (p *VuePlugin) getTypeScriptTemplates() map[string]string {
+	return map[string]string{
+		"src/main.ts":                                vue.ViteTsMainFile,
+		"src/App.vue":                                vue.ViteTsAppFile,
+		"src/components/HelloWorld.vue":              vue.ViteHelloWordVue,
+		"src/components/WelcomeItem.vue":             vue.ViteTsWelcomeItem,
+		"src/components/TheWelcome.vue":              vue.ViteTsTheWelcome,
+		"src/components/icons/IconDocumentation.vue": vue.ViteIconDocumentation,
+		"src/components/icons/IconTooling.vue":       vue.ViteIconTooling,
+		"src/components/icons/IconEcosystem.vue":     vue.ViteIconEcoSystem,
+		"src/components/icons/IconCommunity.vue":     vue.ViteIconCommunity,
+		"src/components/icons/IconSupport.vue":       vue.ViteIconSupport,
+		"src/router/index.ts":                        vue.ViteTsRouter,
+		"src/views/HomeView.vue":                     vue.ViteTsHomeView,
+		"src/views/AboutView.vue":                    vue.ViteTsAboutView,
+		"src/assets/main.css":                        vue.ViteTsMainCss,
+		"src/assets/base.css":                        vue.ViteTsBaseCss,
+	}
 }
 
 func (p *VuePlugin) generateConfigFiles(ctx *tilocontext.ExecutionContext) error {
-	// Generate tsconfig.json
-	tsConfig := `{
-  "extends": "@vue/tsconfig/tsconfig.dom.json",
-  "include": [
-    "env.d.ts",
-    "src/**/*",
-    "src/**/*.vue",
-    "vite.config.*",
-    "vitest.config.*"
-  ],
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  }
-}`
+	language := ctx.Variables["language"].(string)
+	isJS := language == "js"
 
-	// Generate env.d.ts
-	envDts := `/// <reference types="vite/client" />`
-
-	// Generate index.html
-	indexHtml := `<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <link rel="icon" href="/favicon.ico">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>` + ctx.Config.ProjectName + `</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
-  </body>
-</html>`
-
-	// Write config files
-	configs := map[string]string{
-		"tsconfig.json": tsConfig,
-		"src/env.d.ts":  envDts,
-		"index.html":    indexHtml,
+	var configs map[string]string
+	if isJS {
+		configs = p.getJavaScriptConfigs()
+	} else {
+		configs = p.getTypeScriptConfigs()
 	}
 
-	for path, content := range configs {
-		fullPath := filepath.Join(ctx.ProjectPath, path)
+	// Write all config files
+	for relativePath, content := range configs {
+		fullPath := filepath.Join(ctx.ProjectPath, relativePath)
 		if err := utils.WriteFile(fullPath, content); err != nil {
 			return err
 		}
 	}
 
 	return nil
+}
+
+func (p *VuePlugin) getJavaScriptConfigs() map[string]string {
+	return map[string]string{
+		"index.html": vue.ViteJsIndexHtml,
+		".env":       common.Env,
+	}
+}
+
+func (p *VuePlugin) getTypeScriptConfigs() map[string]string {
+	return map[string]string{
+		"tsconfig.json": vue.ViteTsConfig,
+		"src/env.d.ts":  vue.ViteEnvDTs,
+		"index.html":    vue.ViteTsIndexHtml,
+		".env":          common.Env,
+	}
 }
