+++ internal/templates/vue/vite-common.go
@@ -1,0 +1,77 @@
+package vue
+
+const (
+	ViteHelloWordVue = `<script setup lang="ts">
+defineProps<{
+  msg: string
+}>()
+</script>
+
+<template>
+  <div class="greetings">
+    <h1 class="green">{{ msg }}</h1>
+    <h3>
+      You've successfully created a project with
+      <a href="https://vitejs.dev/" target="_blank" rel="noopener">Vite</a> +
+      <a href="https://vuejs.org/" target="_blank" rel="noopener">Vue 3</a> +
+      <a href="https://www.typescriptlang.org/" target="_blank" rel="noopener">TypeScript</a>.
+    </h3>
+  </div>
+</template>
+
+<style scoped>
+h1 {
+  font-weight: 500;
+  font-size: 2.6rem;
+  position: relative;
+  top: -10px;
+}
+
+h3 {
+  font-size: 1.2rem;
+}
+
+.greetings h1,
+.greetings h3 {
+  text-align: center;
+}
+
+@media (min-width: 1024px) {
+  .greetings h1,
+  .greetings h3 {
+    text-align: left;
+  }
+}
+</style>`
+
+	ViteIconDocumentation = `<template>
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="17" fill="currentColor">
+    <path d="M11 2.253a1 1 0 1 0-2 0h2zm-2 13a1 1 0 1 0 2 0H9zM9 2v13h2V2H9z"/>
+    <path d="M5 3a2 2 0 0 0-2 2v6a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V5a2 2 0 0 0-2-2H5zM3 5a2 2 0 0 1 2-2h6a2 2 0 0 1 2 2v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V5z"/>
+  </svg>
+</template>`
+
+	ViteIconTooling = `<template>
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor">
+    <path d="M8 4a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v4H8V4zM6 8h8v8a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2V8z"/>
+  </svg>
+</template>`
+
+	ViteIconEcoSystem = `<template>
+  <svg xmlns="http://www.w3.org/2000/svg" width="18" height="20" fill="currentColor">
+    <path d="M9 0a9 9 0 1 0 9 9A9 9 0 0 0 9 0zm4.5 7.5h-9v-3h9v3zm0 1.5v3h-9v-3h9z"/>
+  </svg>
+</template>`
+
+	ViteIconCommunity = `<template>
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor">
+    <path d="M15 4a1 1 0 1 0 0 2V4zM2 9a1 1 0 0 0 0 2V9zm13-5a1 1 0 1 0 0 2V4zM5 9a1 1 0 0 0 0 2V9z"/>
+  </svg>
+</template>`
+
+	ViteIconSupport = `<template>
+<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor">
+  <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5zM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3zM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2z"/>
+</svg>
+</template>`
+)
